+++ src/main/java/org/opensearch/knn/common/KNNConstants.java
@@ -101,4 +101,9 @@ public class KNNConstants {
 
     // API Constants
     public static final String CLEAR_CACHE = "clear_cache";
+
+    // Filtered Search Constants
+    // Please refer this github issue for more details for choosing this value:
+    // https://github.com/opensearch-project/k-NN/issues/1049#issuecomment-1694741092
+    public static int MAX_DISTANCE_COMPUTATIONS = 2048000;
 }